[refactor] renambe clientId to lfClientId (compared to tcpClientId) in message header. And change type to fixed size (uint64_t)

+++ src/base/network/HookLambdaFunction.hpp
@@ -26,7 +26,7 @@ class HookLambdaFunction : public Hook
 	public:
 		HookLambdaFunction(HookLambdaDef function);
 		virtual ~HookLambdaFunction(void);
-		virtual LibfabricActionResult onMessage(LibfabricConnection * connection, int lfClientId, size_t msgBufferId, LibfabricMessage * clientMessage) override;
+		virtual LibfabricActionResult onMessage(LibfabricConnection * connection, uint64_t lfClientId, size_t msgBufferId, LibfabricMessage * clientMessage) override;
 	private:
 		/** Store the lambda function to be called when recieving a message. **/
 		HookLambdaDef function;
